change: aggiunta loading screen e label per current dataset
