+++ 3Dataviz/frontend/src/components/UI/UI.css
@@ -14,6 +14,18 @@
     background-position: center center;
 }
 
+#current-dataset {
+    position: absolute;
+    top: 0;
+    left: 50%;
+    transform: translateX(-50%);
+    color: white;
+    font-weight: 700;
+    font-size: 1.2rem;
+    background-color: black;
+    padding: .8rem;
+}
+
 button {
     cursor: pointer;
     border-style: none;
