+++ 3Dataviz/frontend/src/components/UI/UI.tsx
@@ -3,12 +3,14 @@ import DataTable from "./DataTable/DataTable";
 import "./UI.css";
 
 type UIProps = {
+  datasetName: string,
   resetCamera: () => void;
 };
 
-function UI({ resetCamera }: UIProps) {
+function UI({ datasetName,resetCamera }: UIProps) {
   return (
     <>
+      <p id="current-dataset" data-testid="current-dataset">{datasetName}</p>
       <DataTable />
       <Options />
       <button
